pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-game-builder"
version = "1.0.0"
description = "Agentic Game-Builder AI — converts NL game ideas into playable browser games"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "litellm>=1.30.0",
    "instructor>=1.0.0",
    "pydantic>=2.0.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
    "playwright>=1.40.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]
production = [
    "prometheus-client>=0.19.0",
    "redis>=5.0.0",
]

[project.scripts]
game-builder = "app.main:app"

[tool.setuptools.packages.find]
include = ["app*", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
